+++ avian3d/CHANGELOG.md
@@ -10,6 +10,9 @@ NOTES:
 * Avian used to be named bevy_xpbd. The old bevy-tnua-xpbd changelog can be seen [here](https://github.com/idanarye/bevy-tnua/blob/3cba881c8825633a8d8bdca1fe30e54500e655b8/xpbd3d/CHANGELOG.md).
 
 ## [Unreleased]
+### Fixed
+- When using `TnuaAvian2dSensorShape`/`TnuaAvian3dSensorShape`, rotate the
+  shape according to changes in gravity direction.
 
 ## 0.5.0 - 2025-05-10
 ### Changed