+++ physics-integration-layer/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `cast_shape_rotation` field for `TnuaProximitySensor`.
 
 ## 0.8.0 - 2025-05-10
 ### Changed